requirements.txt
docling
einops
fastapi
jupyter
langchain
langchain-community
langchain-core
langchain-docling
langchain-google-genai
langchain-openai
langchain-text-splitters
openai
python-dotenv
sentence_transformers
tqdm
weaviate_client